requirements.txt
streamlit
ultralytics
torch
opencv-python
pillow
pytesseract
python-dotenv
openai
numpy
pandas
